refactor(dedupe): extract toggleInArray and useViewMenuFocusGuard from view menus
RuleViewMenu and SessionViewMenu shared two identical code blocks:
- The generic toggleInArray<T> helper (array add/remove without duplicates)
- The focus-guard pattern (changedRef + handleCloseAutoFocus + apply wrapper)
  that defers Radix focus-return to the caller via onApplied

Extracted:
- src/utils/arrayUtils.ts: pure toggleInArray<T> utility
- src/hooks/useViewMenuFocusGuard.ts: generic hook encapsulating the
  focus-guard pattern; returns { apply, handleCloseAutoFocus }

Both call sites updated (RuleViewMenu.tsx, SessionViewMenu.tsx).
The ViewMenuResetItem component was intentionally not extracted: the
reset item uses distinct i18n keys (ruleViewResetLabel vs
sessionViewResetLabel) and different testId strategies, so extraction
would add prop-passing friction with no structural benefit.
